requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-extensions==2.2.9
django-filter==2.2.0
django-graphql-auth==0.3.10
django-graphql-jwt==0.3.0
django-restframework==0.0.1
graphene==2.1.8
graphene-django==2.9.1
graphene-subscriptions==1.0.2
graphql-core==2.3.1
graphql-relay==2.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
mysqlclient==1.4.6
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydotplus==2.0.2
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytz==2019.3
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
zope.interface==5.1.0
